Creating new tree structure for test and adapting to reusable app changes
