+++ runtests.py
@@ -16,7 +16,6 @@ INSTALLED_APPS = [
     'rest_framework',	
     'rest_framework.authtoken',	
     'on_demand',
-    'tests'
 ],	
 
 MIDDLEWARE = [	
@@ -47,34 +46,22 @@ TEMPLATES = [
     },	
 ],
 
-WSGI_APPLICATION = 'on_demand.wsgi.application',	
-
-
-# Database	
-# https://docs.djangoproject.com/en/2.0/ref/settings/#databases	
-
 DATABASES = {	
     'default': {	
-        'ENGINE': 'django.db.backends.mysql', 	
+        'ENGINE': 'django.db.backends.sqlite3',
         'NAME': 'on_demand',	
         'USER': 'root',	
         'PASSWORD': '',	
         'HOST': 'localhost',	
-        'PORT': '3306'	
+        'PORT': '3306'
     }	
 }
 
 )
 
-try:
-    # Django < 1.8
-    from django.test.simple import DjangoTestSuiteRunner
-    test_runner = DjangoTestSuiteRunner(verbosity=2)
-except ImportError:
-    # Django >= 1.8
-    django.setup()
-    from django.test.runner import DiscoverRunner
-    test_runner = DiscoverRunner(verbosity=2)
+django.setup()
+from django.test.runner import DiscoverRunner
+test_runner = DiscoverRunner(verbosity=2)
 
 failures = test_runner.run_tests(['.'])
 if failures:
